+++ include/starter_choose.h
@@ -4,8 +4,8 @@
 extern const u16 gBirchBagGrassPal[2][16];
 extern const u32 gBirchBagTilemap[];
 extern const u32 gBirchGrassTilemap[];
-extern const u32 gBirchHelpGfx[];
-extern const u32 gPokeballSelection_Gfx[];
+extern const u8 gBirchHelpGfx[];
+extern const u8 gPokeballSelection_Gfx[];
 
 u16 GetStarterPokemon(u16 chosenStarterId);
 void CB2_ChooseStarter(void);
